+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - If there are two values entered in one row or column, only count the wrong value as a single mistake.
+- When reavealing the value of a cell, the possible numbers of this cell are now cleared.
 
 ### Security